pyproject.toml
[project]
name = "noteeli"
version = "0.1.0"
description = "Uproszczony webowy notatnik Markdown inspirowany Obsidianem."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "authlib>=1.3.1",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "itsdangerous>=2.2.0",
  "mako>=1.3.5",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "uvicorn app.main:app --reload"
test = "pytest tests/test_workspace_service.py"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
